refactor(BaseController, ManageTable): streamline code by removing unused properties and methods, and introduce TableItem trait for improved item handling and formatting
